Start a TT resize only after search finished.
As noticed in the forum, a crash in extract_ponder_from_tt could result
if hash size is set before the ponder move is printed. While it is
arguably
a GUI issue (but it got me on the cli), it is easy to avoid this issue.

Closes #1856

No functional change.
master
@vondele
@snicolet
vondele authored